edu-visualizer/requirements.txt
streamlit>=1.32.0
openai>=1.14.0
google-genai>=1.11.0
huggingface_hub>=0.26.0
python-dotenv>=1.0.0
Pillow>=10.2.0
requests>=2.31.0
